divide CheckoutFeatures.class
