+++ src/test/resources/features/checkout/payment.feature
@@ -1,26 +1,26 @@
 Feature: Select payment option
 
-  Background:
-    Given logged customer is on home page
-
-  Scenario: User selects pay by bank wire option
-    When user select first item from homepage and added it into cart
-    And he confirms every step to make order and pay by bank wire
-    Then payment by wire is made
-
-  Scenario: User selects pay by check option
-    When user select first item from homepage and added it into cart
-    And he confirms every step to make order and pay by check
-    Then payment by check is made
-
-  Scenario: User selects pay by bank wire option and next changes the payment method to pay by check option
-    When user select first item from homepage and added it into cart
-    And he confirms every step to make order and pay by bank wire
-    And he return to the payment selection and chooses payment by check
-    Then payment by check is made
-
-  Scenario: User selects pay by check option and next changes the payment method to pay by bank wire option
-    When user select first item from homepage and added it into cart
-    And he confirms every step to make order and pay by check
-    And he return to the payment selection and chooses payment by bank wire
-    Then payment by wire is made
+#  Background:
+#    Given logged customer is on home page
+#
+#  Scenario: User selects pay by bank wire option
+#    When user select first item from homepage and added it into cart
+#    And he confirms every step to make order and pay by bank wire
+#    Then payment by wire is made
+#
+#  Scenario: User selects pay by check option
+#    When user select first item from homepage and added it into cart
+#    And he confirms every step to make order and pay by check
+#    Then payment by check is made
+#
+#  Scenario: User selects pay by bank wire option and next changes the payment method to pay by check option
+#    When user select first item from homepage and added it into cart
+#    And he confirms every step to make order and pay by bank wire
+#    And he return to the payment selection and chooses payment by check
+#    Then payment by check is made
+#
+#  Scenario: User selects pay by check option and next changes the payment method to pay by bank wire option
+#    When user select first item from homepage and added it into cart
+#    And he confirms every step to make order and pay by check
+#    And he return to the payment selection and chooses payment by bank wire
+#    Then payment by wire is made
